modules/vision.py
import google.generativeai as genai
import os
from dotenv import load_dotenv
from PIL import Image
import json

load_dotenv()

def analyze_image(image):
    """
    Analyzes an image using Google's Gemini API to detect spoilage.
    Returns a dictionary with risk score (1-10), description, and markers.
    """
    api_key = os.getenv("GEMINI_API_KEY")
    if not api_key:
        return {"error": "API Key not found. Please set GEMINI_API_KEY in .env"}

    genai.configure(api_key=api_key)
    
    # List of models to try in order
    # Trying with 'models/' prefix and standard aliases
    models_to_try = [
        'models/gemini-2.5-flash',
        'models/gemini-2.5-pro',
        'models/gemini-2.0-flash',
        'models/gemini-flash-latest',
        'models/gemini-1.5-flash',
        'models/gemini-1.5-pro',
        'gemini-1.5-flash', # Try without prefix too
    ]

    prompt = """
    You are an agricultural expert. Analyze this image of produce (fruit/vegetable).
    1. Identify the produce.
    2. Estimate a Spoilage Risk Score from 1 to 10 (1 = Fresh/New, 10 = Rotten/Bad).
    3. List specific visual markers of decay or freshness.
    4. Provide a short analysis.

    Return the response strictly as valid JSON with the following structure:
    {
        "produce_name": "string",
        "spoilage_score": integer (1-10),
        "visual_markers": ["marker1", "marker2"],
        "analysis": "string description"
    }
    Do not include markdown formatting (```json ... ```) in the response, just the raw JSON string.
    """

    errors = []

    for model_name in models_to_try:
        try:
            print(f"Trying model: {model_name}...")
            model = genai.GenerativeModel(model_name)
            response = model.generate_content([prompt, image])
            
            # If we get here, it worked
            text = response.text.strip()
            if text.startswith("```json"):
                text = text[7:]
            if text.endswith("```"):
                text = text[:-3]
                
            result = json.loads(text)
            result["model_used"] = model_name # Add debug info
            return result

        except Exception as e:
            error_str = str(e)
            print(f"Model {model_name} failed: {error_str}")
            if "429" in error_str:
                errors.append(f"{model_name}: QUOTA EXCEEDED (429)")
            elif "404" in error_str:
                errors.append(f"{model_name}: NOT FOUND (404)")
            else:
                errors.append(f"{model_name}: {error_str}")
            continue
    
    # --- FALLBACK SIMULATION MODE ---
    # If all API calls failed (likely due to Quota/404), return a MOCK result
    # so the app doesn't crash during the hackathon.
    
    import random
    print("All API models failed. Switching to Simulation Mode.")
    
    # Generate mock score based on random chance for demo
    mock_score = random.randint(3, 8)
    
    mock_files = {
        "produce_name": "Tomato/Vegetable (Simulated)",
        "spoilage_score": mock_score,
        "visual_markers": ["Discoloration (Simulated)", "Soft spots (Simulated)"] if mock_score > 5 else ["Firm skin", "Bright color"],
        "analysis": "⚠️ API Quota Exceeded / Model Unavailable. Using SIMULATION MODE for demo. The produce appears to be analyzed based on heuristic patterns.",
        "model_used": "SIMULATION_MODE (Fallback)"
    }
    
    return mock_files
