requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
pillow
numpy
requests
